Update equilibrium

`update_equilibrium()` updates the equilibrium to a counterfactual situation with new trade costs and/or other changes.

update_equilibrium(
  initial_conditions = NULL,
  tariff = NULL,
  tariff_new = NULL,
  ntb = NULL,
  ntb_new = NULL,
  ntb_change = NULL,
  export_subsidy = NULL,
  export_subsidy_new = NULL,
  consumption_share = NULL,
  factor_share = NULL,
  intermediate_share = NULL,
  trade_elasticity = NULL,
  trade_share = NULL,
  value_added = NULL,
  trade_deficit = NULL,
  max_iterations = 1000,
  tolerance = 1e-04,
  vfactor = 0.1,
  verbose = T
)

Arguments

initial_conditions
Required list of prepared initial conditions, e.g. for GTAP 10.

tariff
Optional data.table of alternative initial tariffs (\(\tau\)), dimensions: origin x destination x sector.

tariff_new
Optional data.table of new tariffs (\(\tau'\)), dimensions: origin x destination x sector.

ntb
Optional data.table of alternative initial non-tariff barriers (\(\kappa\)), dimensions: origin x destination x sector.

ntb_new
Optional data.table of new non-tariff barriers (\(\kappa'\)), dimensions: origin x destination x sector.

ntb_change
Optional data.table of relative change to new non-tariff barriers (\(\kappa' / \kappa\)), dimensions: origin x destination x sector.

export_subsidy
Optional data.table of alternative initial export subsidies (\(\zeta\)), dimensions: origin x destination x sector.

export_subsidy_new
Optional data.table of new export subsidies (\(\zeta'\)), dimensions: origin x destination x sector.

consumption_share
Optional data.table of alternative elasticities of substitution or consumption shares (\(\alpha\)), dimensions: destination x sector.

factor_share
Optional data.table of alternative input factor shares (\(\beta\)), dimensions: country x sector.

intermediate_share
Optional data.table of alternative input-output coefficients (\(\gamma\)), dimensions: country x input_sector x output_sector.

trade_elasticity
Optional data.table of alternative trade elasticities (\(\theta\)), dimensions: sector.

trade_share
Optional data.table of alternative initial trade matrix (\(\pi\)), dimensions: origin x destination x sector.

value_added
Optional data.table of alternative value added (\(w * L\)), dimensions: country.

trade_deficit
Optional data.table of alternative trade balances (\(D\)), dimensions: country.

max_iterations
maximum number of iterations.

tolerance
Tolerance value for iterations.

vfactor
Regulates the "speed" of the adjustment of wages.

verbose
How talkative should the function be?

Value

List of list of data.tables with initial_conditions, model_scenarios and equilibrium_new.